QuestionDao, QuestionnaireDao, ThemeDao add relationship

+++ dao/src/main/java/org/testApp/ConnectUtils/AutoIncrementCompressor.java
@@ -6,7 +6,7 @@ import java.sql.Statement;
 
 public class AutoIncrementCompressor {
 
-    public static void compressionTable(String tableName, int countRows){
+    public static void compressionTable(String tableName, long countRows){
         String query = "ALTER TABLE " + tableName + " auto_increment = " + countRows;
         try(Connection connection = MySQLConnector.getConnection();
             Statement statement = connection.createStatement()){
